Verified that the data loading of the betas is correct.
The model can be overfit on a small subset of the training data
relatively quick and easily.
Adding anti-overfitting methods such as dropout, l2, seem to be
the cause of the weird 1-constant-output on the validation set.
It seems that this goes away with more epochs. Still need to verify
this.