Used Ian's code to generate a directory holding betas and captions.
Currently working on a dataloader for this directory to batch
feed (beta, caption) pairs to the NN model.

I want to create pairs (beta, caption) that minimize the look-up
required for loading the data during the batch_loadign process.
Currently, I have (nsd_key, trial, caption), but the problem is
when I get to the batch_loader I need to do a lot of regex over the
betas folder to get the correct beta.npy file.
It would be better to store (beta_path, caption) as those could be
loaded directly and also don't use much memory.