Copied Ian's pipeline over to Attempt Four.
Created a pandas df/csv with the NSD keys for subject02.
Also re-created the averaged betas and saved those to betas_averaged
in the /huge/.../data directory.

Next I want to take a quick look at the pipeline for reading the data
into the model. Particularly, add the use of the csv file for reading
in NSD keys